Refactor imports and logging in multiple files for consistency and clarity. Updated README.md to simplify bot token description. Enhanced error logging format across handlers for improved debugging. Consolidated model imports in database and service modules for better organization.
