+++ app/middlewares/auth_middleware.py
@@ -1,16 +1,15 @@
 """Auth middleware module for the app."""
 import logging
+from datetime import datetime
 from typing import Any, Awaitable, Callable, Dict
 
-from sqlalchemy import func
 from aiogram import BaseMiddleware
 from aiogram.types import Update
 
 from database.crud import CRUD
-from database.models.user import User
+from database.models import User
 
 
-logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 
@@ -71,7 +70,7 @@ class AuthMiddleware(BaseMiddleware):
 
         # Update user data if it exists
         else:
-            user_data['updated_at'] = func.now()
+            user_data['updated_at'] = datetime.now()
             db_user = await user_crud.update(db_user, **user_data)
             logger.info("User %s updated in the database.", user.id)
