+++ app/utils.py
@@ -1,6 +1,6 @@
 """Utils for the bot."""
 from database.crud import CRUD
-from database.models.search_history import SearchHistory
+from database.models import SearchHistory
 from service.data import Song
