+++ database/models/__init__.py
@@ -1,0 +1,5 @@
+"""Database models."""
+from .user import User
+from .search_history import SearchHistory
+
+__all__ = ['User', 'SearchHistory']
