+++ service/__init__.py
@@ -1,0 +1,5 @@
+"""Service module."""
+from .core import Music
+from .data import Song
+
+__all__ = ['Music', 'Song']
